backend/services/identity_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
django-cors-headers==4.4.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory_boy==3.3.1
Faker==33.3.1
